Some optimizations, using getPropertyValue() instead of getDamage()
