+++ src/main/java/cn/nukkit/block/BlockRedstoneRepeaterUnpowered.java
@@ -1,9 +1,6 @@
 package cn.nukkit.block;
 
 import cn.nukkit.blockstate.BlockState;
-import cn.nukkit.item.Item;
-import cn.nukkit.item.ItemRedstoneRepeater;
-import cn.nukkit.math.BlockFace;
 
 /**
  * @author CreeperFace
@@ -27,26 +24,6 @@ public class BlockRedstoneRepeaterUnpowered extends BlockRedstoneRepeater {
     }
 
     @Override
-    public BlockFace getFacing() {
-        return BlockFace.fromHorizontalIndex(getDamage());
-    }
-
-    @Override
-    protected boolean isAlternateInput(Block block) {
-        return isDiode(block);
-    }
-
-    @Override
-    public Item toItem() {
-        return new ItemRedstoneRepeater();
-    }
-
-    @Override
-    protected int getDelay() {
-        return (1 + getPropertyValue(REPEATER_DELAY)) * 2;
-    }
-
-    @Override
     protected Block getPowered() {
         return BlockState.of(BlockID.POWERED_REPEATER, getCurrentState().getDataStorage()).getBlock();
     }
@@ -55,10 +32,4 @@ public class BlockRedstoneRepeaterUnpowered extends BlockRedstoneRepeater {
     protected Block getUnpowered() {
         return this;
     }
-
-    @Override
-    public boolean isLocked() {
-        return this.getPowerOnSides() > 0;
-    }
-
 }
